center_of_mass {tspmeta}	R Documentation

Return the center of all cities of a TSP instance.

Description

Return the center of all cities of a TSP instance.

Usage

center_of_mass(instance)

Arguments

instance
[tsp_instance]
TSP instance.

Value

[numeric(2)] Center of all cities of the TSP instance.

[Package tspmeta version 1.1 Index]