feature_centroid {tspmeta}	R Documentation

Centroid features.

Description

Includes the coordinates of the mean coordinates of the the point cloud and the statistics of the distances of all cities from it.

Usage

feature_centroid(x)

Arguments

x
[tsp_instance]
TSP instance.

Value

[list].

[Package tspmeta version 1.1 Index]