feature_chull {tspmeta}	R Documentation

Convex hull features.

Description

Determines the area of the convex hull and the ratio of the cities which lie on the convex hull in the euklidean space.

Usage

feature_chull(x)

Arguments

x
[tsp_instance]
TSP instance.

Value

[list].

[Package tspmeta version 1.1 Index]