feature_cluster {tspmeta}	R Documentation

Cluster features.

Description

Determines the number of clusters and the mean distances from all cities in a cluster to its centroid.

Usage

feature_cluster(x, epsilon)

Arguments

x
[tsp_instance]
TSP instance.

epsilon
[numeric(1)]
Probability in [0,1]. Used to compute the reachability distance for the underlying dbscan clustering algorithm.

Value

[list].

[Package tspmeta version 1.1 Index]