greedy_point_matching {tspmeta}	R Documentation

Greedy point matching

Description

Pairs of cities are matched in a greedy fashion for morphing, first the closest pair w.r.t. euclidean distance, then the clostest pair of the remaining cities, and so on.

Usage

greedy_point_matching(x, y)

Arguments

x
[tsp_instance]
First TSP instance

y
[tsp_instance]
Second TSP instance

Value

[matrix] Numeric matrix of point indices with shortest distance.

[Package tspmeta version 1.1 Index]