normalization_angle {tspmeta}	R Documentation

Calculate rotation angle such that the main axis through the cities is aligned with the X axis.

Description

Calculate rotation angle such that the main axis through the cities is aligned with the X axis.

Usage

normalization_angle(instance)

Arguments

instance
[tsp_instance]
TSP instance.

Value

[numeric(1)]

[Package tspmeta version 1.1 Index]