random_instance {tspmeta}	R Documentation

Generates a random TSP instance by scattering random points in a hypercube.

Description

Generates a random TSP instance by scattering random points in a hypercube.

Usage

random_instance(size, d = 2, lower = 0, upper = 1)

Arguments

size
[integer(1)]
Number of cities.

d
[integer(1)]
Space dimensionality, e.g. 2D. Default is 2D.

lower
[numeric(1)]
Lower box constraint for hypercube. Default is 0.

upper
[numeric(1)]
upper box constraint for hypercube. Default is 1.

Value

[tsp_instance].

[Package tspmeta version 1.1 Index]