read_tsplib_instance {tspmeta}	R Documentation

Read in a TSPLIB style Traveling Salesman Problem from a file.

Description

The current state of the parser does not understand all variants of the TSPLIB format. Much effort has been spent making the parser as robust as possible. It will stop as soon as it sees input it cannot handle.

Usage

read_tsplib_instance(path)

Arguments

path
[character(1)]
Character string containing path to file in TSPLIB format.

Value

[tsp_instance].

[Package tspmeta version 1.1 Index]