read_tsplib_instances {tspmeta}	R Documentation

Read in multiple TSPLIB style Traveling Salesman Problems from a directory.

Description

Read in multiple TSPLIB style Traveling Salesman Problems from a directory.

Usage

read_tsplib_instances(path, pattern = "*.tsp", max_size = 1000,
  use_names = TRUE, on_no_coords = "stop")

Arguments

path
[character(1)]
Character string containing path to file in TSPLIB format.

pattern
[character(1)]
Pattern of files under path that are considered as instances.

max_size
[numeric(1)]
Upper bound for instance size (i.e. number of cities). Only applicable, if instance size is contained in file name. Default value ist 1000.

use_names
[logical(1)]
Use base names of files as names of instances in returned list.

on_no_coords
[character(1)]
How to handle instances which do not have any coordinates. Possible values are, “stop” and “warn” which either stop or raise a warning respectivly.

Value

A list List of tsp_instance objects.

[Package tspmeta version 1.1 Index]