read_tsplib_tour {tspmeta}	R Documentation

Read in a TSPLIB style Traveling Salesman Problem tour from a file

Description

Read in a TSPLIB style Traveling Salesman Problem tour from a file

Usage

read_tsplib_tour(path)

Arguments

path
[character(1)]
Filename of file containing a TSP tour.

Value

[TOUR] TOUR object from package TSP, containing order of cities, tour length and method name that generated this solution.

[Package tspmeta version 1.1 Index]