tsp_generation_ea {tspmeta}	R Documentation

TSP generating EA.

Description

TSP generating EA.

Usage

tsp_generation_ea(fitness_function, pop_size = 30L, inst_size = 50L,
  generations = 100L, time_limit = 30L, uniform_mutation_rate,
  normal_mutation_rate, normal_mutation_sd, cells_round = 100L, rnd = TRUE,
  ...)

Arguments

fitness_function
[function(x, ...)]
Fitness function used to judge the fitness of a TSP instance. x is a numeric matrix with 2 columns, containing the coordinates of a TSP instance.

pop_size
[integer(1)]
Number of TSP instances maintained in each population. Default is 30.

inst_size
[integer(1)]
Number of cities of each TSP instance. Default is 50.

generations
[integer(1)]
Number of generations. Default is 100L.

time_limit
[integer(1)]
Time limit in seconds. Default is 30.

uniform_mutation_rate
[numeric(1)]
Mutation probability in uniform mutation (in [0,1]).

normal_mutation_rate
[numeric(1)]
Mutation probability in normal mutation (in [0,1])

normal_mutation_sd
[numeric(1)]
Standard deviation of normal noise in normal mutation

cells_round
[numeric(1)]
Grid resolution for rounding Default is 100.

rnd
[logical(1)]
Round the coordinates before normal mutation. Default is TRUE.

...
[any]
Not used.

Value

[list] List containing best individual form the last population, its fitness value, the genrational fitness and the last population. Default is 50.

[Package tspmeta version 1.1 Index]